requirements.txt
yfinance
requests
beautifulsoup4
pandas
numpy
python-dotenv
pytest
